Backend-Legacy/requirements.txt
# AeroGuard Backend - Production Dependencies
# Tested with Python 3.11.x
# Use: pip install -r requirements.txt

# Core Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Flask-JWT-Extended==4.6.0

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Configuration & Environment
python-dotenv==1.0.0
click==8.1.7

# Data Processing & Analysis
# pandas 2.2.3 is the latest stable with pre-built wheels for Python 3.11 on Linux
pandas==2.2.3
numpy==1.26.4
# statsmodels 0.14.6 satisfies pmdarima>=0.14.5 requirement
statsmodels==0.14.6

# Machine Learning & Forecasting
scikit-learn==1.4.2
pmdarima==2.0.4
xgboost==2.0.3
# tensorflow 2.15.0 is the last release with pre-built Linux wheels for Python 3.11
tensorflow==2.15.0
keras==2.15.0
joblib==1.3.2
ml-dtypes==0.2.0
tensorboard==2.15.2

# HTTP & API
requests==2.31.0
typing-extensions>=4.9.0
protobuf>=3.20.2,<5.0.0
google-generativeai==0.3.2

# Production Server
gunicorn==21.2.0
